Fix bug to parse A tag in text SPAN content
Nexted tag in span.text content was not properly parsed. Sometimes,
A tag can be included in span.text content, but in many cases, we do not
want to include URL info.

This patch enables this BOT to properly remove a tag content from
span.text content, and get text even if A tag is included in the center
of the text.